+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fixed custom tasks suffixing
 - Fixed: Task `:test` uses this output of task `:prepareSandbox` without declaring an explicit or implicit dependency [#1609](../../issues/1609)
 - ExtractorTransformer: Exclude only `Applications` symlink
+- SandboxAware: inherit sandbox directory from producer
 
 ## [2.0.0-beta3] - 2024-05-18